requirements.txt
python-dotenv==1.0.1
tweepy==4.14.0
pylint==3.2.7
black==24.10.0
typer==0.15.1
selenium==4.27.1